refactor: move outside atoms
