+++ packages/shared/components/atoms/index.js
@@ -1,11 +1,3 @@
-export * from './badges'
-export * from './boxes'
-export * from './buttons'
-export * from './iconButtons'
-export * from './labels'
-export * from './pills'
-export * from './tiles'
-
 export { default as AmountBox } from './AmountBox.svelte'
 export { default as BalanceSummaryRow } from './BalanceSummaryRow.svelte'
 export { default as BalanceSummarySection } from './BalanceSummarySection.svelte'
