+++ packages/shared/components/badges/NetworkIconBadge.svelte
