+++ packages/shared/components/badges/NotificationBadge.svelte
