+++ packages/shared/components/badges/StrongholdBadge.svelte
