+++ packages/shared/components/badges/VerificationBadge.svelte
