+++ packages/shared/components/boxes/CopyableBox.svelte
