+++ packages/shared/components/boxes/KeyValueBox.svelte
