+++ packages/shared/components/index.js
@@ -34,3 +34,10 @@ export * from './modals'
 export * from './molecules'
 export * from './organisms'
 export * from './types'
+export * from './badges'
+export * from './boxes'
+export * from './buttons'
+export * from './iconButtons'
+export * from './labels'
+export * from './pills'
+export * from './tiles'
