+++ packages/shared/components/labels/TogglableAssetBalanceLabel.svelte
