+++ packages/shared/components/pills/NetworkStatusPill.svelte
