+++ packages/shared/components/tiles/ShimmerClaimingAccountTile.svelte
